RavenDB-13584 PR Comments
